pyproject.toml
[project]
name = "conv-fin-agent"
version = "0.1.0"
description = "A conversational agent for financial data analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dataclasses-json>=0.6.7",
    "langchain>=0.3.23",
    "langchain-ollama>=0.3.1",
    "langgraph>=0.3.30",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "ruff>=0.11.5",
    "tqdm>=4.67.1",
]
